requirements.txt
langchain
python-dotenv
langchain-openai
ipykernel
pypdf
langchain-community
langchain-core
beautifulsoup4
faiss-cpu
streamlit